requirements.txt
flask==2.3.3
werkzeug==3.0.1
requests==2.26.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
python-dotenv==0.19.0
gunicorn==20.1.0
python-jose==3.3.0
pycryptodome==3.19.0
python-magic==0.4.27
pydantic==2.0.0
google-generativeai==0.4.0
yt-dlp
